+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Disallow uppercase letters in HTTP URL host component.
 - Compare origins instead of hostnames when verifying fetched non-portable object.
 - Make validation of `instance_uri` configuration parameter more strict.
+- Store actor's shared inbox.
 
 ### Fixed